requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.9.0
scenedetect>=0.6.3
faiss-cpu>=1.7.4
open_clip_torch>=2.24.0
tqdm>=4.66.0

